Squashed commit of the following:
commit eb6f0b875c795b982f40b2fe5e2591e19513b5b4
Author: luyoyu <xingctd@gmail.com>
Date:   Mon Aug 5 02:50:31 2024 +0800

    优化重构垃圾代码，约1000行

commit 904965dbddd7ea1a20fd0df7d16f0a2a866bc631
Author: luyoyu <xingctd@gmail.com>
Date:   Sun Aug 4 23:59:53 2024 +0800

    完啦 搞坏了

+++ src/shared/pet_base.cpp
@@ -350,8 +350,8 @@ SItemID CreatePetItem(SItemID item, CItemManager* pItemMgr)
 	SetPetHpAndMp(item,pPet);
 
 	item.params[PET_PARAM_STATE] = 0;
-	int count = 0;
-	int skillCnt = PET_PARAM_PASSIVE_1;
+	size_t count = 0;
+    size_t skillCnt = PET_PARAM_PASSIVE_1;
 	for (int i=PET_PARAM_PASSIVE_1;i<=PET_PARAM_PASSIVE_6;i++)
 	{
 		if (count < pPet->vPassive.size() && count < pPet->vPassiveRate.size()  && GetProb(pPet->vPassiveRate[count]))
